Refactor CLI module for better testability
[noissue]
